load all primi statics
